moved ProcessPluginDefinition and constant interfaces to bpe package
