+++ dsf-bpe/dsf-bpe-process-base/src/main/java/org/highmed/dsf/bpe/ProcessPluginDefinition.java
@@ -1,4 +1,4 @@
-package org.highmed.dsf;
+package org.highmed.dsf.bpe;
 
 import java.util.Collections;
 import java.util.List;
