+++ dsf-bpe/dsf-bpe-process-feasibility/src/main/java/org/highmed/dsf/bpe/FeasibilityProcessPluginDefinition.java
@@ -5,7 +5,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Stream;
 
-import org.highmed.dsf.ProcessPluginDefinition;
 import org.highmed.dsf.bpe.spring.config.FeasibilityConfig;
 import org.highmed.dsf.bpe.spring.config.FeasibilitySerializerConfig;
 import org.highmed.dsf.fhir.resources.AbstractResource;
