+++ dsf-bpe/dsf-bpe-process-feasibility/src/test/java/org/highmed/dsf/fhir/profile/TaskProfileTest.java
@@ -6,32 +6,32 @@ import static org.highmed.dsf.bpe.ConstantsBase.CODESYSTEM_HIGHMED_BPMN_VALUE_CO
 import static org.highmed.dsf.bpe.ConstantsBase.CODESYSTEM_HIGHMED_BPMN_VALUE_ERROR;
 import static org.highmed.dsf.bpe.ConstantsBase.CODESYSTEM_HIGHMED_BPMN_VALUE_MESSAGE_NAME;
 import static org.highmed.dsf.bpe.ConstantsBase.ORGANIZATION_IDENTIFIER_SYSTEM;
-import static org.highmed.dsf.bpe.variables.ConstantsFeasibility.CODESYSTEM_HIGHMED_FEASIBILITY;
-import static org.highmed.dsf.bpe.variables.ConstantsFeasibility.CODESYSTEM_HIGHMED_FEASIBILITY_VALUE_BLOOM_FILTER_CONFIG;
-import static org.highmed.dsf.bpe.variables.ConstantsFeasibility.CODESYSTEM_HIGHMED_FEASIBILITY_VALUE_MULTI_MEDIC_RESULT;
-import static org.highmed.dsf.bpe.variables.ConstantsFeasibility.CODESYSTEM_HIGHMED_FEASIBILITY_VALUE_NEEDS_CONSENT_CHECK;
-import static org.highmed.dsf.bpe.variables.ConstantsFeasibility.CODESYSTEM_HIGHMED_FEASIBILITY_VALUE_NEEDS_RECORD_LINKAGE;
-import static org.highmed.dsf.bpe.variables.ConstantsFeasibility.CODESYSTEM_HIGHMED_FEASIBILITY_VALUE_PARTICIPATING_MEDICS_COUNT;
-import static org.highmed.dsf.bpe.variables.ConstantsFeasibility.CODESYSTEM_HIGHMED_FEASIBILITY_VALUE_PARTICIPATING_MEDIC_CORRELATION_KEY;
-import static org.highmed.dsf.bpe.variables.ConstantsFeasibility.CODESYSTEM_HIGHMED_FEASIBILITY_VALUE_RESEARCH_STUDY_REFERENCE;
-import static org.highmed.dsf.bpe.variables.ConstantsFeasibility.CODESYSTEM_HIGHMED_FEASIBILITY_VALUE_SINGLE_MEDIC_RESULT;
-import static org.highmed.dsf.bpe.variables.ConstantsFeasibility.CODESYSTEM_HIGHMED_FEASIBILITY_VALUE_SINGLE_MEDIC_RESULT_REFERENCE;
-import static org.highmed.dsf.bpe.variables.ConstantsFeasibility.COMPUTE_FEASIBILITY_MESSAGE_NAME;
-import static org.highmed.dsf.bpe.variables.ConstantsFeasibility.COMPUTE_FEASIBILITY_PROCESS_URI_AND_LATEST_VERSION;
-import static org.highmed.dsf.bpe.variables.ConstantsFeasibility.COMPUTE_FEASIBILITY_TASK_PROFILE;
-import static org.highmed.dsf.bpe.variables.ConstantsFeasibility.ERROR_FEASIBILITY_MESSAGE_NAME;
-import static org.highmed.dsf.bpe.variables.ConstantsFeasibility.ERROR_FEASIBILITY_TASK_PROFILE;
-import static org.highmed.dsf.bpe.variables.ConstantsFeasibility.EXECUTE_FEASIBILITY_MESSAGE_NAME;
-import static org.highmed.dsf.bpe.variables.ConstantsFeasibility.EXECUTE_FEASIBILITY_PROCESS_URI_AND_LATEST_VERSION;
-import static org.highmed.dsf.bpe.variables.ConstantsFeasibility.EXECUTE_FEASIBILITY_TASK_PROFILE;
-import static org.highmed.dsf.bpe.variables.ConstantsFeasibility.EXTENSION_GROUP_ID_URI;
-import static org.highmed.dsf.bpe.variables.ConstantsFeasibility.MULTI_MEDIC_RESULT_FEASIBILITY_MESSAGE_NAME;
-import static org.highmed.dsf.bpe.variables.ConstantsFeasibility.MULTI_MEDIC_RESULT_FEASIBILITY_TASK_PROFILE;
-import static org.highmed.dsf.bpe.variables.ConstantsFeasibility.REQUEST_FEASIBILITY_MESSAGE_NAME;
-import static org.highmed.dsf.bpe.variables.ConstantsFeasibility.REQUEST_FEASIBILITY_PROCESS_URI_AND_LATEST_VERSION;
-import static org.highmed.dsf.bpe.variables.ConstantsFeasibility.REQUEST_FEASIBILITY_TASK_PROFILE;
-import static org.highmed.dsf.bpe.variables.ConstantsFeasibility.SINGLE_MEDIC_RESULT_FEASIBILITY_MESSAGE_NAME;
-import static org.highmed.dsf.bpe.variables.ConstantsFeasibility.SINGLE_MEDIC_RESULT_FEASIBILITY_TASK_PROFILE;
+import static org.highmed.dsf.bpe.ConstantsFeasibility.CODESYSTEM_HIGHMED_FEASIBILITY;
+import static org.highmed.dsf.bpe.ConstantsFeasibility.CODESYSTEM_HIGHMED_FEASIBILITY_VALUE_BLOOM_FILTER_CONFIG;
+import static org.highmed.dsf.bpe.ConstantsFeasibility.CODESYSTEM_HIGHMED_FEASIBILITY_VALUE_MULTI_MEDIC_RESULT;
+import static org.highmed.dsf.bpe.ConstantsFeasibility.CODESYSTEM_HIGHMED_FEASIBILITY_VALUE_NEEDS_CONSENT_CHECK;
+import static org.highmed.dsf.bpe.ConstantsFeasibility.CODESYSTEM_HIGHMED_FEASIBILITY_VALUE_NEEDS_RECORD_LINKAGE;
+import static org.highmed.dsf.bpe.ConstantsFeasibility.CODESYSTEM_HIGHMED_FEASIBILITY_VALUE_PARTICIPATING_MEDICS_COUNT;
+import static org.highmed.dsf.bpe.ConstantsFeasibility.CODESYSTEM_HIGHMED_FEASIBILITY_VALUE_PARTICIPATING_MEDIC_CORRELATION_KEY;
+import static org.highmed.dsf.bpe.ConstantsFeasibility.CODESYSTEM_HIGHMED_FEASIBILITY_VALUE_RESEARCH_STUDY_REFERENCE;
+import static org.highmed.dsf.bpe.ConstantsFeasibility.CODESYSTEM_HIGHMED_FEASIBILITY_VALUE_SINGLE_MEDIC_RESULT;
+import static org.highmed.dsf.bpe.ConstantsFeasibility.CODESYSTEM_HIGHMED_FEASIBILITY_VALUE_SINGLE_MEDIC_RESULT_REFERENCE;
+import static org.highmed.dsf.bpe.ConstantsFeasibility.COMPUTE_FEASIBILITY_MESSAGE_NAME;
+import static org.highmed.dsf.bpe.ConstantsFeasibility.COMPUTE_FEASIBILITY_PROCESS_URI_AND_LATEST_VERSION;
+import static org.highmed.dsf.bpe.ConstantsFeasibility.COMPUTE_FEASIBILITY_TASK_PROFILE;
+import static org.highmed.dsf.bpe.ConstantsFeasibility.ERROR_FEASIBILITY_MESSAGE_NAME;
+import static org.highmed.dsf.bpe.ConstantsFeasibility.ERROR_FEASIBILITY_TASK_PROFILE;
+import static org.highmed.dsf.bpe.ConstantsFeasibility.EXECUTE_FEASIBILITY_MESSAGE_NAME;
+import static org.highmed.dsf.bpe.ConstantsFeasibility.EXECUTE_FEASIBILITY_PROCESS_URI_AND_LATEST_VERSION;
+import static org.highmed.dsf.bpe.ConstantsFeasibility.EXECUTE_FEASIBILITY_TASK_PROFILE;
+import static org.highmed.dsf.bpe.ConstantsFeasibility.EXTENSION_GROUP_ID_URI;
+import static org.highmed.dsf.bpe.ConstantsFeasibility.MULTI_MEDIC_RESULT_FEASIBILITY_MESSAGE_NAME;
+import static org.highmed.dsf.bpe.ConstantsFeasibility.MULTI_MEDIC_RESULT_FEASIBILITY_TASK_PROFILE;
+import static org.highmed.dsf.bpe.ConstantsFeasibility.REQUEST_FEASIBILITY_MESSAGE_NAME;
+import static org.highmed.dsf.bpe.ConstantsFeasibility.REQUEST_FEASIBILITY_PROCESS_URI_AND_LATEST_VERSION;
+import static org.highmed.dsf.bpe.ConstantsFeasibility.REQUEST_FEASIBILITY_TASK_PROFILE;
+import static org.highmed.dsf.bpe.ConstantsFeasibility.SINGLE_MEDIC_RESULT_FEASIBILITY_MESSAGE_NAME;
+import static org.highmed.dsf.bpe.ConstantsFeasibility.SINGLE_MEDIC_RESULT_FEASIBILITY_TASK_PROFILE;
 import static org.junit.Assert.assertEquals;
 
 import java.nio.charset.StandardCharsets;
