+++ dsf-bpe/dsf-bpe-process-local-services/src/main/java/org/highmed/dsf/bpe/ConstantsLocalServices.java
@@ -1,0 +1,12 @@
+package org.highmed.dsf.bpe;
+
+import static org.highmed.dsf.bpe.ConstantsFeasibility.LOCAL_SERVICES_PROCESS_URI;
+import static org.highmed.dsf.bpe.LocalServicesProcessPluginDefinition.VERSION;
+
+public interface ConstantsLocalServices
+{
+	String LOCAL_SERVICES_TASK_PROFILE = "http://highmed.org/fhir/StructureDefinition/highmed-task-local-services-integration";
+	String LOCAL_SERVICES_PROCESS_URI_AND_LATEST_VERSION = LOCAL_SERVICES_PROCESS_URI + VERSION;
+
+	String LOCAL_SERVICES_MESSAGE_NAME = "localServicesIntegrationMessage";
+}
