+++ dsf-bpe/dsf-bpe-process-ping/src/main/java/org/highmed/dsf/bpe/ConstantsPing.java
@@ -1,6 +1,7 @@
-package org.highmed.dsf.bpe.variables;
+package org.highmed.dsf.bpe;
 
 import static org.highmed.dsf.bpe.ConstantsBase.PROCESS_URI_BASE;
+import static org.highmed.dsf.bpe.PingProcessPluginDefinition.VERSION;
 
 public interface ConstantsPing
 {
@@ -9,13 +10,11 @@ public interface ConstantsPing
 
 	String PING_TASK_PROFILE = "http://highmed.org/fhir/StructureDefinition/highmed-task-ping";
 	String PING_PROCESS_URI = PROCESS_URI_BASE + "ping/";
-	String PING_PROCESS_LATEST_VERSION = "0.4.0";
-	String PING_PROCESS_URI_AND_LATEST_VERSION = PING_PROCESS_URI + PING_PROCESS_LATEST_VERSION;
+	String PING_PROCESS_URI_AND_LATEST_VERSION = PING_PROCESS_URI + VERSION;
 	String PING_MESSAGE_NAME = "pingMessage";
 
 	String PONG_TASK_PROFILE = "http://highmed.org/fhir/StructureDefinition/highmed-task-pong";
 	String PONG_PROCESS_URI = PROCESS_URI_BASE + "pong/";
-	String PONG_PROCESS_LATEST_VERSION = "0.4.0";
-	String PONG_PROCESS_URI_AND_LATEST_VERSION = PONG_PROCESS_URI + PONG_PROCESS_LATEST_VERSION;
+	String PONG_PROCESS_URI_AND_LATEST_VERSION = PONG_PROCESS_URI + VERSION;
 	String PONG_MESSAGE_NAME = "pongMessage";
 }
